# api/openapi.yaml
openapi: 3.0.0
info:
  title: Phogoprint API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /template:
    post:
      summary: Create a new template
      operationId: CreateTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Template"
      responses:
        "201":
          description: Template created successfully
        "404":
          description: No template found
    get:
      summary: List templates
      operationId: ListTemplate
      responses:
        "200":
          description: The template
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
  /template/{id}:
    get:
      summary: Get a single template
      operationId: GetTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Template not found
  /printer:
    post:
      summary: Print an image directly
      operationId: PrintImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentType
                - data
              properties:
                contentType:
                  type: string
                  example: "image/png"
                data:
                  type: string
                  format: binary
      responses:
        "202":
          description: Printer successfully printed
        "422":
          description: Unprintable image
        "503":
          description: Printer unavailable
  /printer/info:
    get:
      summary: Get device info
      operationId: GetPrinterInfo
      responses:
        "200":
          description: The device info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceInfo"
        "503":
          description: Printer unavailable
components:
  schemas:
    DeviceInfo:
      type: object
      required:
        - batteryLevel
        - firmwareVersion
        - state
      properties:
        batteryLevel:
          type: integer
          description: Battery level of device as a percentage
          example: 100
        firmwareVersion:
          type: string
          description: Device specific firmware version string
          example: "3.0.7"
        state:
          $ref: "#/components/schemas/DeviceState"
    DeviceState:
      type: string
      enum: [DISCONNECTED, CONNECTING, READY, BUSY, OUT_OF_PAPER]
    Template:
      type: object
      required:
        - name
        - landscape
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Postage Label
        landscape:
          type: boolean
          example: true
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/TemplateParameter"
        texts:
          type: array
          items:
            $ref: "#/components/schemas/TemplateText"
        images:
          type: array
          items:
            $ref: "#/components/schemas/TemplateImage"
    TemplateImage:
      type: object
      required:
        - image
        - position
        - width
        - height
      properties:
        image:
          type: string
          example: "irjgt34toofw4ig0j0"
          description: base64 image data
        position:
          $ref: "#/components/schemas/Position"
        width:
          type: integer
          example: 100
        height:
          type: integer
          example: 100
    TemplateText:
      type: object
      required:
        - text
        - position
      properties:
        text:
          type: string
          example: "Post to: {{address}}"
        position:
          $ref: "#/components/schemas/Position"
        width:
          type: integer
          example: 100
        height:
          type: integer
          example: 100
    Position:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          example: 10
        y:
          type: integer
          example: 10
    TemplateParameter:
      type: object
      required:
        - name
        - maxLength
      properties:
        name:
          type: string
          example: address
        maxLength:
          type: integer
          example: 100